Release IPPrefixes: support AWS ENI IP Prefix unassignment
Cilium currently does not support releasing IPPrefixes and secondary
IPs when IPPrefix delegation enabled due to which IP starvation is
happening in our Kubernetes clusters. We have to manually delete a
node in order to free unused IPPrefixes.

With this fix, the operator will release those unused IPPrefixes and
secondary IPs to reassign them again in AWS ENI.

Added a no-op stub implementation in Alibaba and Azure Clouds to satisfy the
NodeOperations interface.

Fixes: #32209, #39904

Signed-off-by: Harish Salluri <hsalluri259@gmail.com>